moved visualization data to dashboard_app dir
