+++ stellarisdashboard/cli.py
@@ -12,7 +12,8 @@ import threading
 
 import click
 
-from stellarisdashboard import config, visualization_data
+from stellarisdashboard import config
+from stellarisdashboard.dashboard_app import visualization_data
 from stellarisdashboard.parsing import save_parser, timeline
