+++ stellarisdashboard/dashboard_app/visualization_data.py
